Remove bigquery code; Add filtering capability with col stats enpdoint
fix

add back join statement

wip
