+++ src/components/dataset/data/sidebar/panels/FilterPanel.jsx
@@ -104,9 +104,7 @@ export class FilterPanel extends React.PureComponent {
     dataset: PropTypes.object,
     dispatch: PropTypes.func.isRequired,
     filterData: PropTypes.object,
-    filterStatement: PropTypes.string,
     handleCreateSnapshot: PropTypes.func,
-    joinStatement: PropTypes.string,
     open: PropTypes.bool,
     selected: PropTypes.string,
     table: PropTypes.object,
@@ -199,11 +197,9 @@ export class FilterPanel extends React.PureComponent {
       columns,
       dataset,
       filterData,
-      filterStatement,
       open,
       table,
       token,
-      joinStatement,
       selected,
       handleCreateSnapshot,
       canLink,
@@ -323,8 +319,6 @@ function mapStateToProps(state) {
     dataset: state.datasets.dataset,
     columns: state.query.columns,
     filterData: state.query.filterData,
-    filterStatement: state.query.filterStatement,
-    joinStatement: state.query.joinStatement,
     token: state.user.delegateToken,
   };
 }
